Move body generator for untagged enums to its own module
Cut-paste
